Fix navbar reactivity Switch to table view
