+++ src/lib/logged-in-store.svelte.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const loggedInStore = writable(false);
