+++ src/routes/recipes/+page.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
-	import { Button } from '@sveltestrap/sveltestrap';
-	import CardDeck from '../../components/CardDeck.svelte';
-	import RecipeListItem from '../../components/RecipeListItem.svelte';
-	import type { PageProps } from './$types';
 	import { getDatabaseEntries } from '$lib/models/database';
+	import { Button, ButtonGroup, Icon, Table } from '@sveltestrap/sveltestrap';
+	import type { PageProps } from './$types';
 
 	async function handleDelete(key: string) {
 		const resp = await fetch(`/api/recipe/${key}`, {
@@ -30,8 +28,28 @@
 
 <Button color="primary" href="/recipes/new">New Recipe</Button>
 
-<CardDeck>
-	{#each recipes as [key, recipe] (key)}
-		<RecipeListItem {recipe} {key} onDelete={() => handleDelete(key)} />
-	{/each}
-</CardDeck>
+<Table striped bordered hover>
+	<thead>
+		<tr>
+			<th>Name</th>
+			<th>Actions</th>
+		</tr>
+	</thead>
+	<tbody>
+		{#each recipes as [key, recipe] (key)}
+			<tr>
+				<td>{recipe.name}</td>
+				<td>
+					<ButtonGroup>
+						<Button href={`/recipes/${key}`} color="warning">
+							<Icon name="pencil-square" />
+						</Button>
+						<Button color="danger" on:click={() => handleDelete(key)}>
+							<Icon name="trash" />
+						</Button>
+					</ButtonGroup>
+				</td>
+			</tr>
+		{/each}
+	</tbody>
+</Table>
