+++ src/routes/takeout/+page.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
-	import { Button } from '@sveltestrap/sveltestrap';
-	import CardDeck from '../../components/CardDeck.svelte';
-	import TakeoutListItem from '../../components/TakeoutListItem.svelte';
-	import type { PageProps } from './$types';
 	import { getDatabaseEntries } from '$lib/models/database';
+	import { Button, ButtonGroup, Icon, Table } from '@sveltestrap/sveltestrap';
+	import type { PageProps } from './$types';
 
 	async function handleDelete(key: string) {
 		const resp = await fetch(`/api/takeout/${key}`, {
@@ -33,6 +31,23 @@
 		takeoutsRaw[newTakeout.id] = newTakeout;
 	}
 
+	async function handleEdit(key: string) {
+		const takeout = takeoutsRaw[key];
+		if (!takeout) {
+			console.error('Takeout not found:', key);
+			return;
+		}
+		const newName = window.prompt('Please enter the new name for the takeout:', takeout.name);
+		if (newName && newName.trim() !== '') {
+			takeout.name = newName;
+			await fetch(`/api/takeout/${key}`, {
+				method: 'PUT',
+				headers: { 'Content-Type': 'application/json' },
+				body: JSON.stringify({ name: takeout.name })
+			});
+		}
+	}
+
 	let { data }: PageProps = $props();
 
 	let takeoutsRaw = $state(data.takeouts);
@@ -43,8 +58,28 @@
 
 <Button color="primary" on:click={() => addTakeout()}>New takeout</Button>
 
-<CardDeck>
-	{#each takeouts as [key, takeout] (key)}
-		<TakeoutListItem {takeout} {key} onDelete={() => handleDelete(key)} />
-	{/each}
-</CardDeck>
+<Table striped bordered hover>
+	<thead>
+		<tr>
+			<th>Name</th>
+			<th>Actions</th>
+		</tr>
+	</thead>
+	<tbody>
+		{#each takeouts as [key, takeout] (key)}
+			<tr>
+				<td>{takeout.name}</td>
+				<td>
+					<ButtonGroup>
+						<Button color="warning" on:click={() => handleEdit(key)}>
+							<Icon name="pencil-square" />
+						</Button>
+						<Button color="danger" on:click={() => handleDelete(key)}>
+							<Icon name="trash" />
+						</Button>
+					</ButtonGroup>
+				</td>
+			</tr>
+		{/each}
+	</tbody>
+</Table>
